Refactor the structure of the crash handler
