+++ libraries/networking/src/crash-handler/CrashHandler.cpp
@@ -10,18 +10,33 @@
 //
 
 #include "CrashHandler.h"
-#include "CrashHandlerBackend.h"
+#include "CrashHandlerNone.h"
+#include "CrashHandlerSentry.h"
+
 #include <QFileInfo>
 #include <QCoreApplication>
 #include <QDebug>
+#include <QDir>
+
 #include "LogHandler.h"
+
+#ifdef Q_OS_LINUX
+#include <unistd.h>
+#endif
+
+Q_LOGGING_CATEGORY(crash_handler, "overte.crash_handler")
 
 
 static void crashHandlerLogMessage(QtMsgType type, const QMessageLogContext &context, const QString &msg);
 
 
 CrashHandler& CrashHandler::getInstance() {
-    static CrashHandler sharedInstance;
+#if HAS_SENTRY
+    static CrashHandlerSentry sharedInstance;
+#else
+    static CrashHandlerNone sharedInstance;
+#endif
+
     return sharedInstance;
 }
 
@@ -50,7 +65,7 @@ bool CrashHandler::start() {
         return false;
     }
 
-    auto started = startCrashHandler(_path.toStdString(), _crashUrl.toStdString(), _crashToken.toStdString());
+    auto started = startCrashHandler();
     setStarted(started);
 
     if ( started ) {
@@ -82,7 +97,7 @@ bool CrashHandler::start() {
 }
 
 void CrashHandler::startMonitor(QCoreApplication *app) {
-    startCrashHookMonitor(app);
+   // startCrashHookMonitor(app);
 }
 
 void CrashHandler::setEnabled(bool enabled) {
@@ -132,6 +147,58 @@ void CrashHandler::setAnnotation(const std::string &key, const std::string &valu
     }
 
     setCrashAnnotation(key, value);
+}
+
+
+void CrashHandler::logMessage(QtMsgType type, const QMessageLogContext &context, const QString &msg) {
+    // Some of the logging is very spammy and Sentry seems to have a problem with shutting down
+    // in that case.
+    {
+        std::lock_guard<std::mutex> lock(_logMutex);
+
+        if (msg == _previousMessage) {
+            _repeatCount++;
+            return;
+        }
+
+        if (_repeatCount > 0) {
+            sendLogMessage(QtMsgType::QtWarningMsg, QMessageLogContext(), QString("Previous log message repeated %1 times").arg(_repeatCount));
+        }
+
+        _previousMessage = msg;
+        _repeatCount = 0;
+    }
+
+    sendLogMessage(type, context, msg);
+}
+
+// Locate the full path to the binary's directory
+QString CrashHandler::findBinaryDir() {
+    // Normally we'd just use QCoreApplication::applicationDirPath(), but we can't.
+    // That function needs the QApplication to be created first, and Crashpad is initialized as early as possible,
+    // which is well before QApplication, so that function throws out a warning and returns ".".
+    //
+    // So we must do things the hard way here. In particular this is needed to correctly handle things in AppImage
+    // on Linux. On Windows and MacOS falling back to argv[0] should be fine.
+
+#ifdef Q_OS_LINUX
+    // Find outselves by looking at /proc/<PID>/exe
+    pid_t ourPid = getpid();
+    QString exeLink = QString("/proc/%1/exe").arg(ourPid);
+    qCDebug(crash_handler) << "Looking at" << exeLink;
+
+    QFileInfo exeLinkInfo(exeLink);
+    if (exeLinkInfo.isSymLink()) {
+        QFileInfo exeInfo(exeLinkInfo.symLinkTarget());
+        qCDebug(crash_handler) << "exe symlink points at" << exeInfo;
+        return exeInfo.absoluteDir().absolutePath();
+    } else {
+        qCWarning(crash_handler) << exeLink << "isn't a symlink. /proc not mounted?";
+    }
+
+#endif
+
+    return QString();
 }
